requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
uvicorn
fastapi
together
sse_starlette